algunas actualizaciones de tailwind

+++ database/migrations/2024_11_19_165801_create_plans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table->string('name'); 
-            $table->decimal('price', 8, 2);
+            $table->unsignedInteger('price');
             $table->timestamps();
         });
     }
